Refactor out encryption intent code to separate subclass
