+++ src/index.ts
@@ -19,6 +19,7 @@ export * from "./components/request-factory";
 
 export * from "./components/client-factory";
 export * from "./components/encryption";
+export * from "./components/encrypted-intent";
 export * from "./components/intent";
 export * from "./components/room-link-validator";
 export * from "./components/room-upgrade-handler";
